chore: refactory balances and sales domain to create new domains payments and cashflow
