+++ internal/logger/interface.go
@@ -1,5 +1,6 @@
 package logger
 
+// TODO: Mover para /internal/logger
 type Logger interface {
 	Debugf(format string, args ...interface{})
 	Infof(format string, args ...interface{})
