+++ sales/sales.go
@@ -4,7 +4,7 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/pkg/validator"
+	"github.com/uesleicarvalhoo/sorveteria-tres-estrelas/internal/validator"
 )
 
 type Item struct {
